Separate BMDB and VCDB routes and controllers; moved LLM BMDB query to llms_router
