+++ backend/app/routes/llms_router.py
@@ -1,6 +1,8 @@
 from multiprocessing import process
-
-from fastapi import APIRouter, Request
+from fastapi import APIRouter, Depends, HTTPException, Response
+from typing import List
+from app.schemas.bmdb_schema import BMDBRequestParams
+from app.schemas.vcelldb_schema import BiomodelRequestParams
 import httpx
 import requests
 from app.controllers.llms_controller import (
@@ -11,6 +13,25 @@ from app.controllers.llms_controller import (
 )
 
 router = APIRouter()
+
+# For BioModelsDB search using BioModelsDB API 
+@router.post("/bmdb-search")
+async def search_llm(conversation_history: dict):
+    """
+    Endpoint to query the LLM and execute the necessary tools.
+    Args:
+        conversation_history (dict): The conversation history containing user prompts and responses.
+        database (str): The database to query - bmdb in this case.
+    Returns:
+        dict: The final response after processing the prompt with the tools.
+    """
+
+    print("DEBUG20: BMDB POST: ROUTER")
+    result, bmdbkeys, tool_summary = await get_llm_response(
+        conversation_history.get("conversation_history", []), database="bmdb"
+    )
+    return {"response": result, "bmkeys": bmdbkeys, "tool_summary": tool_summary}
+
 
 
 @router.post("/query")
